refactor: 라이브러리 코드 폴더 구조 수정 - lib 1뎁스 내용을 각 라이브러리 폴더에 추가
