+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
 import { QueryClientProvider } from "@tanstack/react-query";
-import { queryClient } from "@/lib/react-query-client";
+import { queryClient } from "@/lib/tanstack-query/react-query-client";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
